remove and rename files to navigate faster through the directory structure
