+++ src/com/Git/Repository/Filtering/Filters.java
@@ -18,4 +18,6 @@ public class Filters
     {
         return pathname.getName().toLowerCase().endsWith(".txt");
     }
+
+
 }
