+++ src/com/Git/Repository/Generics/Filtering/Filter.java
