--- src/com/Similarity/Similarity.java
@@ -1,17 +1,0 @@
-package com.Similarity;
-
-public class Similarity
-{
-    private Similarity()
-    {
-
-    }
-
-    static
-    {
-        Instance = new Similarity();
-    }
-
-
-    public static Similarity Instance;
-}
