--- src/com/Similarity/laboaratoare/LuceneConstants.java
@@ -1,8 +1,0 @@
-package laboaratoare;
-
-public class LuceneConstants {
-    public static final String CONTENTS = "contents";
-    public static final String FILE_NAME = "filename";
-    public static final String FILE_PATH = "filepath";
-    public static final int MAX_SEARCH = 10;
-}
