--- src/com/Similarity/laboaratoare/Main.java
@@ -1,8 +1,0 @@
-package laboaratoare;
-
-public class Main {
-
-    public static void main(String[] args) {
-
-    }
-}
