--- src/com/Similarity/laboaratoare/TextFileFilter.java
@@ -1,12 +1,0 @@
-package laboaratoare;
-
-import java.io.File;
-import java.io.FileFilter;
-
-public class TextFileFilter implements FileFilter {
-
-    @Override
-    public boolean accept(File pathname) {
-        return pathname.getName().toLowerCase().endsWith(".txt");
-    }
-}
